+++ markwon-core/src/main/java/io/noties/markwon/core/CorePlugin.java
@@ -115,7 +115,20 @@ public class CorePlugin extends AbstractMarkwonPlugin {
     // @since 4.0.0
     private final List<OnTextAddedListener> onTextAddedListeners = new ArrayList<>(0);
 
+    // @since $nap;
+    private boolean hasExplicitMovementMethodPlugin;
+
     protected CorePlugin() {
+    }
+
+    /**
+     * @since $nap;
+     */
+    @SuppressWarnings("UnusedReturnValue")
+    @NonNull
+    public CorePlugin hasExplicitMovementMethodPlugin(boolean hasExplicitMovementMethodPlugin) {
+        this.hasExplicitMovementMethodPlugin = hasExplicitMovementMethodPlugin;
+        return this;
     }
 
     /**
@@ -188,7 +201,8 @@ public class CorePlugin extends AbstractMarkwonPlugin {
         // let's ensure that there is a movement method applied
         // we do it `afterSetText` so any user-defined movement method won't be
         // replaced (it should be done in `beforeSetText` or manually on a TextView)
-        if (textView.getMovementMethod() == null) {
+        // @since $nap; we additionally check if we should apply _implicit_ movement method
+        if (!hasExplicitMovementMethodPlugin && textView.getMovementMethod() == null) {
             textView.setMovementMethod(LinkMovementMethod.getInstance());
         }
     }